+++ public/index.html
@@ -12,7 +12,6 @@
 <body>
     <div id="app">
     </div>
-    <script src="https://cdn.jsdelivr.net/npm/vue@2/dist/vue.js"></script>
-</body>
-
+    	<script src="https://cdn.jsdelivr.net/npm/vue@2/dist/vue.js"></script>
+    </body>
 </html>
