--- src/assets/Services.js
@@ -1,17 +1,0 @@
-export default [
-    {
-        servimage : "https://www.flaticon.com/svg/static/icons/svg/1970/1970073.svg",
-        servtitulo : "Domótica",
-        servdescrip : "Control de iluminación, entretenimiento, seguridad y confort a su alcance.",
-    },
-    {
-        servimage : "https://cdn2.iconfinder.com/data/icons/5g-internet/512/5g-internet-wifi-technology-14-512.png",
-        servtitulo : "Recolección de datos ambientales",
-        servdescrip : "Optimizar el proceso de cultivar la tierra mediante el monitoreo.",
-    },
-    {
-        servimage : "https://www.flaticon.com/svg/static/icons/svg/984/984496.svg",
-        servtitulo : "Visualización de datos",
-        servdescrip : "Representación gráfica de información y datos.",
-    }
-];
