--- src/assets/Team.js
@@ -1,50 +1,0 @@
-var Team = [{
-        codigo: 1,
-        nombre: 'Angela Maria Agudelo M.',
-        descripcion:'',
-        caracteristicas:{
-            universidad:'Universidad Pontificia Bolivariana',
-            hobbie:'Ver series y jugar Flag Football',
-            edad:'24 años',
-            correo:'angela.agudeloa@upb.edu.co',
-        },
-        rol: 'Desarrollador Frontend',
-        image: 'https://scontent.fbog2-1.fna.fbcdn.net/v/t1.0-9/130191100_10221094008033108_6443151654411284691_n.jpg?_nc_cat=100&ccb=2&_nc_sid=730e14&_nc_ohc=ZSuoCEjJOdIAX8v8HPr&_nc_ht=scontent.fbog2-1.fna&oh=f5021ec06719e001e0e4a5ab587ba1d6&oe=5FF2C322'
-},{
-        codigo: 2,
-        nombre: 'Juan Pablo Rodríguez F.',
-        descripcion: '',
-        caracteristicas:{
-            universidad:'Universidad del Cauca',
-            hobbie:'Fútbol Sala',
-            edad:'27 años',
-            correo:'juanpablor@unicauca.edu.co',
-        },
-        rol: 'Desarrollador Backend',
-        image: 'https://scontent.fbog2-1.fna.fbcdn.net/v/t1.0-9/129955345_10221094007873104_2398791372188815673_n.jpg?_nc_cat=102&ccb=2&_nc_sid=730e14&_nc_ohc=Vn-rLiTL3I8AX_z4O0z&_nc_ht=scontent.fbog2-1.fna&oh=55cc0ee83944f198f199cb5644e3d597&oe=5FF0EB8D'
-},{
-        codigo: 3,
-        nombre: 'Brayan Yesid Marín A.',
-        descripcion: '',
-        caracteristicas:{
-            universidad:'Universidad ECCI',
-            hobbie:'Leer, Ver series',
-            edad:'25 años',
-            correo:'brayan.yesid.marin@gmail.com',
-        },
-        rol: '<Analista y Tester de software',
-        image: 'https://scontent.fbog2-1.fna.fbcdn.net/v/t1.0-9/129052269_10221094008073109_4949163993691958678_n.jpg?_nc_cat=101&ccb=2&_nc_sid=730e14&_nc_ohc=7cd8gARsxLcAX-gZTj2&_nc_ht=scontent.fbog2-1.fna&oh=aff0aeb636f750b9dd0d70ede18e1296&oe=5FF02CA2'
-},{
-        codigo: 4,
-        nombre: 'Iván Mateo Bohórquez P.',
-        descripcion:'',
-        caracteristicas:{
-            universidad:'Universidad Piloto de Colombia',
-            hobbie:'Jugar videojuegos',
-            edad:'22 años',
-            correo:'ivanmateobohorquez@gmail.com',
-        },
-        rol: 'Arquitecto de software',
-        image: 'https://scontent.fbog2-2.fna.fbcdn.net/v/t1.0-9/130193078_10221094007913105_7891312505364346317_n.jpg?_nc_cat=105&ccb=2&_nc_sid=730e14&_nc_ohc=9sFQQ1DDCYIAX_8-V2Z&_nc_ht=scontent.fbog2-2.fna&oh=7f58d3509b80bd189d477c9f839b5f4d&oe=5FF041D2'
-}];
-export default Team;
