--- src/assets/index.js
@@ -1,2 +1,0 @@
-export * from './js/data.js'
-
