--- src/assets/js/data.js
@@ -1,58 +1,0 @@
-var services = [
-	{
-		name: 'Dómotica',
-		urlToImage: 'https://www.flaticon.com/svg/static/icons/svg/3378/3378068.svg',
-		description: 'Control de iluminación, entretenimiento, seguridad y confort a su alcance.',
-		rigth: 'https://www.flaticon.com/authors/dighital',
-		author: 'dighital',
-	},{
-		name: 'Recolección de datos ambientales',
-		urlToImage: 'https://www.flaticon.com/svg/static/icons/svg/3377/3377974.svg',
-		description: 'Optimizar el proceso de cultivar la tierra mediante el monitoreo.',
-		rigth: 'https://www.flaticon.com/authors/dighital',
-		author: 'dighital',
-	},{
-		name: 'Visualización de datos',
-		urlToImage: 'https://www.flaticon.com/svg/static/icons/svg/3378/3378153.svg',
-		description: 'Representación gráfica de información y datos.',
-		rigth: 'https://www.flaticon.com/authors/dighital',
-		author: 'dighital',
-	},
-];
-
-var team = [
-	{
-		codigo: 1,
-		nombre: '',
-		descripcion: '',
-		rol: '',
-		image: '',
-		scrum: ''
-	},{
-		codigo: 2,
-		nombre: '',
-		descripcion: '',
-		rol: '',
-		image: '',
-		scrum: ''
-	},{
-		codigo: 3,
-		nombre: '',
-		descripcion: '',
-		rol: '',
-		image: '',
-		scrum: ''
-	},{
-		codigo: 4,
-		nombre: '',
-		descripcion: '',
-		rol: '',
-		image: '',
-		scrum: ''
-	}
-]
-
-export { 
-	services,
-	team,
-}
