--- src/components/ComponenteHeader.vue
@@ -1,17 +1,0 @@
-<template>
-    <div id="header">
-        <div class="fondo-img" id="home">
-				<img src="../static/img/background-min.jpg" alt="moubi" width="100%" height="350px">
-		</div>
-    </div>
-</template>
-
-<script>
-    export default {
-        name: "ComponenteHeader"
-    }
-</script>
-
-<style scoped>
-
-</style>
