+++ src/components/NavHeader.vue
@@ -1,0 +1,17 @@
+<template>
+	<div id="header">
+		<div class="fondo-img" id="home">
+			<img src="../static/img/background-min.jpg" alt="moubi" style="width: 100%;"/>
+		</div>
+	</div>
+</template>
+
+<script>
+    export default {
+        name: 'NavHeader',
+    }
+</script>
+
+<style scoped>
+
+</style>
