+++ src/components/index.js
@@ -1,9 +1,15 @@
 import NewsCard from './NewsCard.vue';
 import TeamCard from './TeamCard.vue';
 import ServiceCard from './ServiceCard.vue';
+import NavFooter from './NavFooter.vue';
+import NavHeader from './NavHeader.vue';
+import Navigation from './Navigation.vue';
 
 export {
 	NewsCard,
 	TeamCard,
 	ServiceCard,
+	NavFooter, 
+	NavHeader,
+	Navigation,
 };
